06-ft_transcendence/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.8
psycopg2-binary==2.9.9
sqlparse==0.4.4
typing_extensions==4.8.0
django-bootstrap-v5==1.0.11
pillow==10.1.0
channels==4.0.0
daphne==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-otp==1.3.0
django-formtools==2.4.1
django-two-factor-auth==1.15.5
qrcode==7.4.2
django-environ==0.11.2
hvac==2.0.0
gunicorn==21.2.0
Twisted[tls,http2]
django-cors-headers==4.3.1
